#49 - Checking installed package version before uinstalling.
Until now Uninstall process from Browser tab was trying to remove package that was found.
This could caused problems as that found version could has different content.
After this commit selected package on Browser tab can have all buttons disabled:
- Any version of package is installed, so Install command is not available.
- Selected version can't be uninstalled (or reinstalled) because other one is installed.
This should not be the final solution as it's UX is not good.